+++ src/main/java/life/qbic/omero/BasicOMEROClient.java
@@ -1,6 +1,25 @@
 package life.qbic.omero;
 
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
+import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
+import javax.imageio.ImageIO;
+import omero.ServerError;
+import omero.api.IAdminPrx;
+import omero.api.RenderingEnginePrx;
+import omero.api.ThumbnailStorePrx;
 import omero.gateway.Gateway;
 import omero.gateway.LoginCredentials;
 import omero.gateway.SecurityContext;
@@ -9,22 +28,25 @@ import omero.gateway.exception.DSOutOfServiceException;
 import omero.gateway.facility.BrowseFacility;
 import omero.gateway.facility.DataManagerFacility;
 import omero.gateway.facility.MetadataFacility;
-import omero.ServerError;
-import omero.api.RenderingEnginePrx;
-import omero.api.ThumbnailStorePrx;
-import omero.gateway.model.*;
+import omero.gateway.model.AnnotationData;
+import omero.gateway.model.ChannelData;
+import omero.gateway.model.DatasetData;
+import omero.gateway.model.ExperimenterData;
+import omero.gateway.model.FileAnnotationData;
+import omero.gateway.model.ImageData;
+import omero.gateway.model.MapAnnotationData;
+import omero.gateway.model.PixelsData;
+import omero.gateway.model.ProjectData;
 import omero.log.SimpleLogger;
-import omero.model.*;
+import omero.model.Dataset;
+import omero.model.DatasetI;
+import omero.model.IObject;
+import omero.model.NamedValue;
+import omero.model.Project;
+import omero.model.ProjectDatasetLink;
+import omero.model.ProjectDatasetLinkI;
+import omero.model.ProjectI;
 import omero.romio.PlaneDef;
-import java.util.*;
-import java.util.concurrent.ExecutionException;
-import javax.imageio.ImageIO;
-import java.awt.image.BufferedImage;
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
-
-
 
 /////////////////////////////////////////////////////
 
@@ -34,7 +56,7 @@ public class BasicOMEROClient {
   /** Reference to the gateway. */
   private Gateway gateway;
   /** The security context. */
-  private SecurityContext ctx;
+  private SecurityContext securityContext;
 
   private String hostname;
   private int port;
@@ -46,6 +68,7 @@ public class BasicOMEROClient {
 
   private HashMap<Long, String> projectMap;
   private HashMap<Long, Set<DatasetData>> datasetMap;
+  private String sessionUuid;
 
   public BasicOMEROClient(String username, String password, String hostname, int port) {
 
@@ -59,11 +82,14 @@ public class BasicOMEROClient {
   }
 
   /**
+   * Connects the the omero gateway. Uses the experimenter data to update properties of this omero client instance.
    *
    * @param hostname
    * @param port
    * @param username
    * @param password
+   * @see Gateway
+   * @see ome.system.EventContext
    */
   private void connect(String hostname, int port, String username, String password) {
 
@@ -80,11 +106,15 @@ public class BasicOMEROClient {
 
     try {
       ExperimenterData user = gateway.connect(cred);
-      this.ctx = new SecurityContext(user.getGroupId());
-
+      this.securityContext = new SecurityContext(user.getGroupId());
       this.sessionId = gateway.getSessionId(user);
+      IAdminPrx gatewayAdminService = gateway.getAdminService(securityContext);
+      this.sessionUuid = gatewayAdminService.getEventContext().sessionUuid;
+
     } catch (DSOutOfServiceException dsOutOfServiceException) {
       throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
+    } catch (ServerError serverError) {
+      throw new RuntimeException("Omero store interaction failed.", serverError);
     }
   }
 
@@ -124,7 +154,7 @@ public class BasicOMEROClient {
 
     try {
       BrowseFacility browse = gateway.getFacility(BrowseFacility.class);
-      image = browse.getImage(this.ctx, imageID);
+      image = browse.getImage(this.securityContext, imageID);
     } catch (DSOutOfServiceException dsOutOfServiceException) {
       throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
     } catch (ExecutionException executionException) {
@@ -139,7 +169,7 @@ public class BasicOMEROClient {
 
     try {
       MetadataFacility metadata = gateway.getFacility(MetadataFacility.class);
-      annotations = metadata.getAnnotations(ctx, image, types, null);
+      annotations = metadata.getAnnotations(securityContext, image, types, null);
     } catch (DSOutOfServiceException dsOutOfServiceException) {
       throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
     } catch (ExecutionException executionException) {
@@ -169,7 +199,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
     RenderingEnginePrx proxy;
     ByteArrayInputStream stream;
     try {
-      proxy = gateway.getRenderingService(ctx, pixelsId);
+      proxy = gateway.getRenderingService(securityContext, pixelsId);
       proxy.lookupPixels(pixelsId);
       if (!(proxy.lookupRenderingDef(pixelsId))) {
         proxy.resetDefaultSettings(true);
@@ -213,11 +243,10 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
   }
 
   private SecurityContext getContext() {
-    return ctx;
+    return securityContext;
   }
 
   public String getSessionId() {
-
     return this.sessionId;
   }
 
@@ -244,10 +273,10 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
     String downloadLinkAddress;
     try {
       BrowseFacility browse = gateway.getFacility(BrowseFacility.class);
-      ImageData image = browse.getImage(this.ctx, imageID);
+      ImageData image = browse.getImage(this.securityContext, imageID);
       disconnect();
       if (image.getFormat() != null) {
-        downloadLinkAddress = "http://" + hostname + "/omero/webgateway/archived_files/download/" + imageID + "?server=1&bsession=" + this.sessionId;
+        downloadLinkAddress = "http://" + hostname + "/omero/webgateway/archived_files/download/" + imageID + "?server=1&bsession=" + sessionUuid;
       } else {
         throw new IllegalArgumentException("No image format given. Image is not available for download.");
       }
@@ -280,7 +309,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
     try {
 
       BrowseFacility browse = gateway.getFacility(BrowseFacility.class);
-      Collection<ProjectData> projects = browse.getProjects(ctx);
+      Collection<ProjectData> projects = browse.getProjects(securityContext);
 
       Iterator<ProjectData> i = projects.iterator();
       ProjectData project;
@@ -294,8 +323,12 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
         this.datasetMap.put(id, project.getDatasets());
       }
 
-    } catch (Exception e) {
-      System.out.println(e);
+    } catch (DSOutOfServiceException dsOutOfServiceException) {
+      throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
+    } catch (ExecutionException executionException) {
+      throw new RuntimeException("Task aborted unexpectedly.", executionException);
+    } catch (DSAccessException dsAccessException) {
+      throw new RuntimeException("Could not pull data from the omero server.", dsAccessException);
     }
 
     return this.projectMap;
@@ -308,7 +341,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
     try {
 
       BrowseFacility browse = gateway.getFacility(BrowseFacility.class);
-      Collection<ProjectData> projects = browse.getProjects(ctx);
+      Collection<ProjectData> projects = browse.getProjects(securityContext);
 
       Iterator<ProjectData> i = projects.iterator();
       ProjectData project;
@@ -322,11 +355,13 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
 
           break;
         }
-
       }
-
-    } catch (Exception e) {
-      System.out.println(e);
+    } catch (DSOutOfServiceException dsOutOfServiceException) {
+      throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
+    } catch (ExecutionException executionException) {
+      throw new RuntimeException("Task aborted unexpectedly.", executionException);
+    } catch (DSAccessException dsAccessException) {
+      throw new RuntimeException("Could not pull data from the omero server.", dsAccessException);
     }
 
     return projectInfo;
@@ -371,7 +406,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
 
     IObject r;
     try {
-      r = dm.saveAndReturnObject(this.ctx, proj);
+      r = dm.saveAndReturnObject(this.securityContext, proj);
     } catch (DSOutOfServiceException dsOutOfServiceException) {
       throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
     } catch (DSAccessException dsAccessException) {
@@ -400,7 +435,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
 
     IObject r = null;
     try {
-      r = dm.saveAndReturnObject(this.ctx, link);
+      r = dm.saveAndReturnObject(this.securityContext, link);
     } catch (DSOutOfServiceException dsOutOfServiceException) {
       throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
     } catch (DSAccessException dsAccessException) {
@@ -427,7 +462,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
 
     try {
       DataManagerFacility fac = gateway.getFacility(DataManagerFacility.class);
-      fac.attachAnnotation(ctx, data, new ProjectData(new ProjectI(projectId, false)));
+      fac.attachAnnotation(securityContext, data, new ProjectData(new ProjectI(projectId, false)));
     } catch (DSOutOfServiceException dsOutOfServiceException) {
       throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
     } catch (ExecutionException executionException) {
@@ -448,14 +483,15 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
 
     data.setNameSpace(MapAnnotationData.NS_CLIENT_CREATED);
     try {
-
       DataManagerFacility fac = gateway.getFacility(DataManagerFacility.class);
-      fac.attachAnnotation(ctx, data, new DatasetData(new DatasetI(datasetId, false)));
-
-    } catch (Exception e) {
-      System.out.println(e);
+      fac.attachAnnotation(securityContext, data, new DatasetData(new DatasetI(datasetId, false)));
+    } catch (DSOutOfServiceException dsOutOfServiceException) {
+      throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
+    } catch (ExecutionException executionException) {
+      throw new RuntimeException("Task aborted unexpectedly.", executionException);
+    } catch (DSAccessException dsAccessException) {
+      throw new RuntimeException("Could not pull data from the omero server.", dsAccessException);
     }
-
   }
 
   public HashMap<Long, String> getImages(long datasetId) {
@@ -463,20 +499,21 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
     HashMap<Long, String> imageList = new HashMap<Long, String>();
 
     try {
-
       BrowseFacility browse = this.gateway.getFacility(BrowseFacility.class);
-      Collection<ImageData> images = browse.getImagesForDatasets(ctx, Arrays.asList(datasetId));
+      Collection<ImageData> images = browse.getImagesForDatasets(securityContext, Arrays.asList(datasetId));
 
       Iterator<ImageData> j = images.iterator();
       ImageData image;
       while (j.hasNext()) {
         image = j.next();
         imageList.put(image.getId(), image.getName());
-
       }
-
-    } catch (Exception e) {
-      System.out.println(e);
+    } catch (DSOutOfServiceException dsOutOfServiceException) {
+      throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
+    } catch (ExecutionException executionException) {
+      throw new RuntimeException("Task aborted unexpectedly.", executionException);
+    } catch (DSAccessException dsAccessException) {
+      throw new RuntimeException("Could not pull data from the omero server.", dsAccessException);
     }
 
     return imageList;
@@ -490,7 +527,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
 
       BrowseFacility browse = this.gateway.getFacility(BrowseFacility.class);
       Collection<ImageData> images =
-          browse.getImagesForDatasets(this.ctx, Arrays.asList(datasetId));
+          browse.getImagesForDatasets(this.securityContext, Arrays.asList(datasetId));
 
       Iterator<ImageData> j = images.iterator();
       ImageData image = null;
@@ -520,7 +557,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
         MetadataFacility mdf = gateway.getFacility(MetadataFacility.class);
 
         String channelNamesString = "";
-        List<ChannelData> data = mdf.getChannelData(ctx, imageId);
+        List<ChannelData> data = mdf.getChannelData(securityContext, imageId);
         for (ChannelData c : data) {
           channelNamesString = channelNamesString + c.getName() + ", ";
         }
@@ -531,11 +568,32 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
 
 
 
-    } catch (Exception e) {
-      System.out.println(e);
+    } catch (DSOutOfServiceException dsOutOfServiceException) {
+      throw new RuntimeException("Error while accessing omero service: broken connection, expired session or not logged in", dsOutOfServiceException);
+    } catch (ExecutionException executionException) {
+      throw new RuntimeException("Task aborted unexpectedly.", executionException);
+    } catch (DSAccessException dsAccessException) {
+      throw new RuntimeException("Could not pull data from the omero server.", dsAccessException);
     }
 
     return imageInfo;
+  }
+
+  /**
+   * This method returns a http address at which the given image can be viewed using the omero web client.
+   * @param imageId the omero id of a selected image
+   * @return an address at which the given image can be viewed using the omero web client
+   */
+  public String composeImageDetailAddress(long imageId) {
+    int serverId = 1;
+    return "http://"
+            + hostname
+            + "/omero/webclient/img_detail/"
+            + imageId
+            + "/?server="
+            + serverId
+            + "&bsession="
+            + sessionUuid;
   }
 
   /**
@@ -552,7 +610,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
     try {
       BrowseFacility browse = this.gateway.getFacility(BrowseFacility.class);
       Collection<ImageData> images =
-          browse.getImagesForDatasets(this.ctx, Collections.singletonList(datasetId));
+          browse.getImagesForDatasets(this.securityContext, Collections.singletonList(datasetId));
 
       Iterator<ImageData> j = images.iterator();
       ImageData image = null;
@@ -563,7 +621,7 @@ return (annotations != null) ? annotations.stream().map(annotationData -> (T) an
         }
       }
 
-      store = this.gateway.getThumbnailService(ctx);
+      store = this.gateway.getThumbnailService(securityContext);
 
       PixelsData pixels = Objects.requireNonNull(image).getDefaultPixels();
       store.setPixelsId(pixels.getId());